SaveMnemonicModal in ReceiveSendButtons component
